+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Renamed global method helpers service provider
 - Renamed collection service provider
 - `transposeWithKeys` now automatically guesses the row header names and allows you to override them
+- Renamed constant helper trait to `Constants` (because of the lack of a better name)
 
 ### Removed